requirements.txt
flask
appwrite
python-dotenv